refactor(radio): rename component for docs sake
