+++ src/components/Radio/index.js
@@ -1,2 +1,2 @@
-export { default as MRadio } from './src/RadioFormItem.vue';
+export { default as MRadio } from './src/Radio.vue';
 export { default as MRadioControl } from './src/RadioControl.vue';
